feature - 짝꿍 해제 #35
