+++ service-internal-api/src/main/java/com/dongrami/user/ui/UserController.java
@@ -4,10 +4,12 @@ import com.dongrami.common.ApiResponse;
 import com.dongrami.user.application.UserService;
 import com.dongrami.user.domain.UserEntity;
 import com.dongrami.user.dto.request.RequestDeactivation;
-import com.dongrami.user.dto.request.RequestInviteCode;
 import com.dongrami.user.dto.request.RequestUpdateNotification;
 import com.dongrami.user.dto.request.RequestUpdateProfileInfo;
-import com.dongrami.user.dto.response.*;
+import com.dongrami.user.dto.response.ResponseCountTodoAndDiary;
+import com.dongrami.user.dto.response.ResponseProfileInfo;
+import com.dongrami.user.dto.response.ResponseUserInfo;
+import com.dongrami.user.dto.response.ResponseUserNotification;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -85,42 +87,6 @@ public class UserController implements UserControllerInterface {
         ResponseCountTodoAndDiary response = userService.countTodoAndDiary(principal.getUsername());
 
         return ResponseEntity.ok(ApiResponse.success(response));
-    }
-
-    /**
-     * 짝꿍 초대코드 조회
-     */
-    @GetMapping("/users/invite")
-    public ResponseEntity<?> getInviteCode(@AuthenticationPrincipal User principal) {
-        String inviteCode = userService.getInviteCode(principal.getUsername());
-
-        return ResponseEntity.ok(
-                ApiResponse.success(new ResponseInviteCode(inviteCode))
-        );
-    }
-
-    /**
-     * 짝꿍 초대코드 입력
-     */
-    @PostMapping("/users/invite")
-    public ResponseEntity<?> updateInviteCode(@AuthenticationPrincipal User principal, @Valid @RequestBody RequestInviteCode request) {
-        userService.updateInviteCode(principal.getUsername(), request);
-
-        return ResponseEntity.ok(
-                ApiResponse.success()
-        );
-    }
-
-    /**
-     * 짝꿍 정보 조회
-     */
-    @GetMapping("/users/pair")
-    public ResponseEntity<?> getPairUserInfo(@AuthenticationPrincipal User principal) {
-        ResponsePairUserInfo response = userService.getPairUserInfo(principal.getUsername());
-
-        return ResponseEntity.ok(
-                ApiResponse.success(response)
-        );
     }
 
     /**
